Work around what is probably a bug in f95
Complaint was "Error: Symbol ‘j’ at (1) is declared in a BLOCK DATA
program unit but is not in a COMMON block [-Werror=unused-variable]".
However j was just used in implied do loops in data statements.

Fix by including j in the COMMON block (bletch).

Also transpose tstdat, so it's filled in column major order.  (Testing
out alternative fix, to fill the whole array.  But that fails because
there are too many continuations lines.)